requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anthropic==0.40.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
APScheduler==3.10.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asn1crypto==1.5.1
asyncio-redis
attrs==23.2.0
Authlib==1.3.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
boto3==1.34.139
botocore==1.34.139
cachetools==5.3.3
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
cryptography==42.0.8
cycler==0.12.1
dataclasses-json==0.6.7
dataclasses-json-speakeasy==0.5.11
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
diskcache==5.6.3
distro==1.8.0
dnspython==2.6.1
docstring_parser==0.16
effdet==0.4.1
fastapi
fastembed
filelock==3.15.4
filetype==1.2.0
FlashRank==0.2.6
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.1
geocoder
geopy
google-ai-generativelanguage==0.6.10
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-cloud==0.34.0
google-cloud-aiplatform==1.74.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-speech==2.26.0
google-cloud-storage==2.19.0
google-crc32c==1.5.0
google-generativeai==0.8.3
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-health-checking==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
h11==0.14.0
h2==4.1.0
hiredis==2.2.3
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.1.0
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.3.11
langchain-anthropic==0.3.0
langchain-community==0.3.11
langchain-core==0.3.28
langchain-experimental==0.3.3
langchain-google-genai==2.0.7
langchain-google-vertexai==2.0.9
langchain-groq==0.2.1
langchain-ibm==0.3.5
langchain-neo4j==0.1.1
langchain-openai==0.2.12
langchain-qdrant==0.1.4
langchain-text-splitters==0.3.2
langchainhub==0.1.20
langdetect==1.0.9
langgraph==0.2.60
langsmith==0.2.6
layoutparser==0.3.4
lingua-language-detector==2.0.2
llama_cpp_python==0.2.76
llama-index==0.12.19
llama-index-embeddings-fastembed==0.3.0
llama-index-readers-file==0.4.5
llama-index-storage-index-store-redis==0.4.0
llama-index-vector-stores-qdrant==0.4.3
loguru==0.7.2
lomond==0.3.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib
mdurl==0.1.2
mlflow==2.16.2
mmh3==4.1.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.18.1
openai==1.57.4
opencv-python==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-propagator-b3==1.25.0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
ordered-set==4.1.0
orjson==3.10.6
outcome==1.3.0.post0
packaging==24.1
pandas==2.1.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.1
pikepdf==9.0.0
pillow==10.4.0
pillow_heif==0.17.0
platformdirs==4.2.2
portalocker==2.10.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
proto-plus==1.24.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
PyMuPDF==1.24.10
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==5.1.0
PyPDF2==3.0.1
pypdfium2==4.30.0
PySocks==1.7.1
PyStemmer==2.2.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-logstash==0.4.8
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.12.2
rapidfuzz==3.9.4
redis==5.0.8
redisvl==0.3.7
rejson
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.3
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.22.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==75.1.0
shapely==2.0.4
shellingham==1.5.4
siconv==0.1.2.9
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
snowflake-connector-python==3.11.0
snowflake-sqlalchemy==1.0.8
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.8.7
SpeechRecognition==3.10.4
SQLAlchemy==2.0.31
starlette==0.37.2
structlog==24.2.0
sympy==1.12.1
tabulate==0.9.0
tavily-python==0.3.3
tenacity==8.4.2
textblob==0.18.0.post0
threadpoolctl==3.5.0
tiktoken==0.7.0
timm==1.0.7
tokenizers==0.19.1
tomlkit==0.12.5
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
types-protobuf==5.27.0.20240626
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
unstructured==0.14.9
unstructured-client==0.23.8
unstructured-inference==0.7.36
unstructured.pytesseract==0.3.12
uritemplate==4.1.1
urllib3==2.1.0
uvicorn==0.30.1
validators==0.28.3
watchfiles==0.22.0
weaviate-client==4.6.5
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto
yarl==1.9.4
zipp==3.19.2
trafilatura
readability-lxml
lxml_html_clean
google-generativeai